+++ src/app/search-module/components/search-page/search-page.component.ts
@@ -1,16 +1,18 @@
-import { Component, ViewEncapsulation } from '@angular/core';
+import { Component, inject, ViewEncapsulation } from '@angular/core';
 import { Title } from '@angular/platform-browser';
-import { APP_PAGE_TITLE } from "../../../app.settings";
-import { LanguageService } from "../../../shared/services/language-service/language.service";
+import { APP_PAGE_TITLE } from '../../../app.settings';
+import { LanguageService } from '../../../shared/services/language-service/language.service';
 import locale from './search.locale.json';
 
 @Component({
   selector: 'app-search-page',
   template: `
     <!--page container-->
-    <div class="py-10 mx-auto max-w-[920px]">
-      <h2 class="font-bold text-4xl text-center">{{ locale[langService.language].FindRecipes }}</h2>
-      <div class="my-8">
+    <div class="py-5 md:py-10 mx-auto max-w-[920px]">
+      <h2 class="font-bold text-3xl md:text-4xl text-center">
+        {{ locale[langService.language].FindRecipes }}
+      </h2>
+      <div class="my-4 md:my-8">
         <p class="text-center mb-2">{{ locale[langService.language].SearchText1 }}</p>
         <!--TODO after click of underline text, focus on element?-->
         <p class="text-center" [innerHTML]="locale[langService.language].SearchText2"></p>
@@ -21,15 +23,13 @@ import locale from './search.locale.json';
       <app-recipe />
     </div>
   `,
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
 })
 export class SearchPageComponent {
   protected readonly locale = locale;
+  protected readonly langService = inject(LanguageService);
 
-  constructor(
-    readonly title: Title,
-    protected readonly langService: LanguageService
-  ) {
-    title.setTitle(APP_PAGE_TITLE.SEARCH[langService.language]);
+  constructor(readonly title: Title) {
+    title.setTitle(APP_PAGE_TITLE.SEARCH[this.langService.language]);
   }
 }