+++ src/app/search-module/services/recipe.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable, Subject, tap } from 'rxjs';
-
 import { environment } from '../../../environments/environment';
 import { ResponseList } from '../../shared/models/response-list.type';
 import { Recipe } from '../models/recipe.interface';
@@ -15,13 +14,16 @@ export class RecipeService {
 
   /**
    * Find recipes based on ingredients
+   * @param token Recaptcha token
    * @param ingredients List of ingredient codenames
    */
-  findRecipes(ingredients: number[]): Observable<ResponseList<Recipe>> {
-    const queryParams = new HttpParams().set('ingredients', JSON.stringify(ingredients));
+  findRecipes(token = '', ingredients: number[]): Observable<ResponseList<Recipe>> {
+    const params = new HttpParams()
+      .set('ingredients', JSON.stringify(ingredients))
+      .set('token', token);
     return this.http
       .get<ResponseList<Recipe>>(`${environment.SERVER_API}/recipe/by-ingredients`, {
-        params: queryParams,
+        params,
       })
       .pipe(tap((recipes) => this.onSearch.next(recipes)));
   }