+++ src/environments/environment.prod.ts
@@ -2,4 +2,5 @@ export const environment = {
   production: true,
   SERVER_ORIGIN: 'https://recipeappexpress-7211.rostiapp.cz',
   SERVER_API: 'https://recipeappexpress-7211.rostiapp.cz/api/v1',
+  RECAPTCHA_KEY: '6Lc7Wc0pAAAAALZDS65zf6T8viyOkrTNqZ-pey5q'
 };