+++ src/environments/environment.ts
@@ -2,4 +2,5 @@ export const environment = {
   production: false,
   SERVER_ORIGIN: 'http://localhost:3000',
   SERVER_API: 'http://localhost:3000/api/v1',
+  RECAPTCHA_KEY: '6Lc7Wc0pAAAAALZDS65zf6T8viyOkrTNqZ-pey5q'
 };