+++ CHANGELOG.md
@@ -10,6 +10,14 @@ Run command for (without commit):
 
 ## Versions
 
+### 1.17.0
+
+- using recaptcha service when search recipes
+
+### 1.16.7
+
+- prevent trigger to submit Search recipe form multiple times
+
 ### 1.16.6
 
 - set title for Error page (404)